Fix assertions linked to the wrong test
Because we rely on log:changed events to flag an assertion ended,
it's possible for a new test to start (resetting currentTest) before
that event has arrived. When it does arrive, it adopts the new test
for its events causing a mismatch in the step grouper. This seems
to be most prevalent when the assertion is in a before() call where
there aren't any active tests when the assertion starts.

Caching the "current test" in the assertion handler when we start
and using that for any subsequent events in the handler should
result in the same behavior for typical assertions and also resolve
the bug because we'll be assigning the step:end for the assertion
to the correct test (even though a new test has started).