Allow uploading directly to S3
It's pretty easy to overwhelm our servers by sending them a ton of CI
recordings at once. It's been happening multiple times a day sometimes
for weeks. In the case of dozens or hundreds of recordings coming from
CI, there is not really any advantage to using a websocket over HTTP,
and HTTP is a *lot* simpler (and we don't have to worry about scale
ever, it just goes straight to AWS).