Allow unknown test metadata versions
Future versions of metadata might be uploaded by an older
versions of the cli which will fail when really it just
doesn't know about the format. Might be better to not
validate existing metadata on upload under the assumption
that it was already validated but this will do the trick
for now.