Fix missing test metadata when step parsing fails
The v2 metadata change caused a regression in which we lost
track of the test title and scope when a step parsing error
occurs which made all the tests show the same default title
of the spec filename.

This splits the test result to test matching into a separate
function so we can safely get the initial test metadata
independently from step matching.