Refine update prompt logic (#359)
The new CLI will prompt the user to update either the CLI itself or the runtime if either is out of date. I added logic initially so that we wouldn't ask more than once a day (because that would be annoying) but I realized that we _should_ probably ask again if _a new_ release came out sooner than a day. (For example, if we do a roll back of the Runtime or something?)

So this commit changes the "should we prompt again?" logic to check both _when_ we last asked and what the last value we suggested updating to was.

---

I also bumped the `package.json` version bump in anticipation of publishing a preview release once this change lands. (Hoping to avoid additional PR review delay.)