Add REPLAY_API_KEY fallback for RECORD_REPLAY_API_KEY (#415)
We weren't doing this everywhere, just most places. I think this change is important to align with the newer version of the flag that we are mentioning in [the instructions in our documentation](https://docs.replay.io/test-runners/playwright/record-your-first-replay)

I also updated a few docs to reference `REPLAY_API_KEY` instead of the legacy `RECORD_REPLAY_API_KEY` variable.