Don't try to group related recordings; just sort reverse chronologically (#442)
Trying to use metadata to group "related" recordings is *flawed* and may result in more confusion for the end user (see the discussion thread on TT-1051). This commit removes that logic. I think the right thing to do, for now, is to remove that workaround and wait until the runtime supports it natively.

Note that CLI output isn't as nice as before in the multi-recording "happy path" scenario:

![Screenshot 2024-05-06 at 2 18 04 PM](https://github.com/replayio/replay-cli/assets/29597/35d83cf9-8d41-43a4-b185-b3d5acb97407)

![Screenshot 2024-05-06 at 2 18 16 PM](https://github.com/replayio/replay-cli/assets/29597/2fc86ae1-c2bd-4618-81ff-8873b54dd2ef)